mistral/lechat_custom_mcp_server/notes/requirements.txt
fastmcp==2.5.1
mistralai==1.5.0
python-dotenv==1.1.0
